Disable Rails/WhereRange
This caused a lot of changes which can be autocorrected, but not safely,
making it more effort and risk to change.

It also changes SQL conditions from `>= ?` `<= ?` to `..` and
`>` `<` to `...`. I really don't think this is an improvement, and it is
harder to spot (and understand) the difference between `..` and `...`.